third_party/blink/web_tests/platform/generic/fast/js/mozilla/strict/15.5.5.1-expected.txt

PASS 'use strict'; var s = str(); s.length = 1; s.length threw exception of type TypeError.
PASS var s = str(); s.length = 1; s.length is 3
PASS true === true
PASS 'use strict'; var s = str(); delete s.length threw exception of type TypeError.
PASS var s = str(); delete s.length is false
PASS true === true
PASS 'use strict'; "foo".length = 1 threw exception of type TypeError.
PASS "foo".length = 1 is 1
PASS true === true
PASS 'use strict'; delete "foo".length threw exception of type TypeError.
PASS delete "foo".length is false
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
